Added FFL and BFL, with new vertex property in element
We now track the front and back vertices of an element.
 They are given in reference to the front edge of the element. For a Lens(), they overlap at z=0, for Space() they are both None. Even for matrix groups, we can actually calculate the front and back vertices fairly easily.

For more complex elements (such as Objective()), we need to set the frontVertex and backVertex manually in the __init__ because we do not model the interfaces explicitly: we model the objective as a perfect thin lens, but we still know where the real vertices are. By keeping track of them, we get EFL, BFL and FFL in all cases.