Complete refactoring of Rays
It appeared clear that Rays should use a list and provide all the methods for computing the histograms when needed.  Since we only simulate millions of rays, it only takes a few megabytes of memory to keep all rays, so that does not affect performance that much.
Also, we will start implementing an "intensity" for Ray() to accelerate the simulations: for histograms, this will be used as a weight instead of having to simulate a number of rays proportionnal to the intensity at that point.  This "wavepacket" strategy is common for Monte Carlo.