Move to multiprocessing brought a htread-safety issue
Fixed a thread-safety issue in traceManyThrough by explictly creating the list of rays before splitting it in separate lists. Lock() are not possible to use with Pool()

We also default to the number of cpus as given by multiprocessing

This could still be optimized because the use of map of large arrays is not optimal and the benefit is not proportionnal to the number of CPUs.