*Major* changes to fix images being drawn out of the display
Added imagingPath.images reference and split drawImages into compute and draw. The displayRange property now checks the computed .images heights so that they are displayed properly. and .createRayTracePlot first computes the images before using the displayRange property. This can be discussed as there might be a better way to keep that reference.