Update testsMaterials.py
I think it is simpler and clearer to use this:
Material.__subclasses__()
to get all the subclasses of Material.  I had made a mistake with my comment originally (it is a function).

We do not want to fail at the tearDown part! The test is over at that point. I rewrote with a local variable.