Mistake on Thorlabs web site
The achromat AC254_030_A at https://www.thorlabs.com/newgrouppage9.cfm?objectgroup_id=120
does not mention the back focal length that is mentionned in the Graph PDF drawing available on the same line.  We take the PDF value.

For AC254-125-A, it is not possible to validate the edge thickness on other drawings.