update stub files (#269)
We must regenerate the stub files for the documentation manually.

The stub files of the documentation (all small .rst files in raytracing/docs/modules) are not regenerated by ReadTheDocs... and I have yet to find a way. For now it's still refreshed by hand by deleting the stub files and typing sphinx-autogen -t docs/_templates docs/reference.rst in a console.