Creation of `utils.py`
The numpy function `isclose` is actualy surprinsingly slow.  I was able to create a smalleer function that is 10x faster.  Since we are using it in `Matrix.__init__`, it is critical to be fast and as it stood, it slowed down the plotting of LaserPath and others significantly.  I created 4 functions:
`isAlmostZero()`, `isAlmostOne()`, `areAlmostEqual()` and `areNotEqual()`