Added a new test to compare with current Doublets in module
The AchromatDoublets were created from the main web page, and the values are given to 1 decimal, however, the actual prescription is with 2 decimals and it does make a difference: AC254_100_A() was off by 0.05 mm and putting the same numbers as the actual Zemax file solved the discrepancy. Now they are identical (BFL, FFL, etc...).

I changed the deisgn wavelength to 0.586 to everyone because I need to find a way to read the WAVM parameter from the file.  Thorlabs list 15 wavelengths, Edmund only one.  I will need to figure this out.