MOved functions from Zemax to Material and utils
Matching materials is a Material function not Zemax.
Moved a few functions out of zemax.py and to utils or materials.
Got a better function for finding materials by name, return an exception if not found to indicate what to do.
This showed an existing error where th Edmund test file had a material unknown to raytracing but did not return an error.  Material added, New tests pass.