Fix non zero length matrices
In the process of making the GRIN, I added a new parameter `dz` to allow one to split the propagation easily for all elements (such as space and GRIN).
This had the side-effect to make illegal the creation of a matrix manually with a non-zero length without overriding `transferMatrix`.
I had to create a new PP1ToPP2 matrix for the objective to propagate from the front prinicpal plane to the back plane.