First Raytrace and Raytraces attempt
It has become necesary to create a class for raytrace and raytraces, especially because of the OpenCL implementation.  CompactRays is a convenient way to keep all rays in one large array. For proper performance, passing around compact representation is essential: the first OpenCL test were impressive, but copying the results back took forever. It has become necessary to create a Raytrace that appears like a list of rays, and Raytraces which is a list of raytrace but using a buffer to retrieve the data.  The shared buffer is a very efficient way of avoiding to copy many objects.